Change: project structure: modes as modules in separate files
